requirements.txt
beautifulsoup4==4.12.2
EbookLib==0.18
openai==1.35.9
PyYAML==6.0.1